refactor: add a stub for non-Windows platforms
