+++ rust/libs/logging/src/lib.rs
@@ -9,7 +9,6 @@ mod capturing_writer;
 mod display_btree_set;
 mod err_with_sources;
 mod event_message_contains_filter;
-#[cfg(windows)]
 pub mod windows_event_log;
 
 use std::sync::Arc;
